+++ src/Links/links.css
@@ -1,9 +1,15 @@
+.links-container {
+    margin: 0;
+    padding: 0;
+}
 .links-item {
     display: flex;
     flex-wrap: no-wrap;
     width: 300px;
     align-items: center;
     justify-content: space-between;
+    align-content: center;
+    align-self: center;
 }
 .link-item > img {
     width: 30px;
